fix computed type

+++ cloudtower/models/datacenter.py
@@ -47,9 +47,9 @@ class Datacenter(object):
         'total_cpu_hz': 'int',
         'total_data_capacity': 'int',
         'total_memory_bytes': 'int',
-        'used_cpu_hz': 'int',
+        'used_cpu_hz': 'float',
         'used_data_space': 'int',
-        'used_memory_bytes': 'int',
+        'used_memory_bytes': 'float',
         'vm_num': 'int'
     }
 
@@ -354,7 +354,7 @@ class Datacenter(object):
 
 
         :return: The used_cpu_hz of this Datacenter.  # noqa: E501
-        :rtype: int
+        :rtype: float
         """
         return self._used_cpu_hz
 
@@ -364,7 +364,7 @@ class Datacenter(object):
 
 
         :param used_cpu_hz: The used_cpu_hz of this Datacenter.  # noqa: E501
-        :type used_cpu_hz: int
+        :type used_cpu_hz: float
         """
 
         self._used_cpu_hz = used_cpu_hz
@@ -396,7 +396,7 @@ class Datacenter(object):
 
 
         :return: The used_memory_bytes of this Datacenter.  # noqa: E501
-        :rtype: int
+        :rtype: float
         """
         return self._used_memory_bytes
 
@@ -406,7 +406,7 @@ class Datacenter(object):
 
 
         :param used_memory_bytes: The used_memory_bytes of this Datacenter.  # noqa: E501
-        :type used_memory_bytes: int
+        :type used_memory_bytes: float
         """
 
         self._used_memory_bytes = used_memory_bytes
